MIGRATION dates should be unique within a document -- it doesn't help us to show dupes

+++ app/models/metadata_date.rb
@@ -7,6 +7,7 @@ class MetadataDate < ActiveRecord::Base
   # Destroy and recreate all of a document's dates, from the text. Save the
   # document after running this method in order to save the dates.
   def self.refresh(document)
+    return false unless document.text
     document.metadata_dates.destroy_all
     DC::Import::DateExtractor.new.extract_dates(document.text).each do |date|
       document.metadata_dates << self.new(:document => document, :date => date)
